added a live preview
What good is a model if you can't see real time updates?  I think live previews are awesome and they are super easy in Angular.  Its something that is not so easy in Meteor immediately without installing some extra plugins.  If it makes the section too long, maybe it can go somewhere else....i'm honestly only at this point in the tutorial :-)